ats/requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
matplotlib
langchain_experimental
sentence_transformers
groq
cassio
beautifulsoup4
langchain-groq
wikipedia
arxiv
langchainhub
PyPDF2